backend/requirements.txt
Django==2.2.19
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
djoser==2.1.0
django-filter==21.1
django-colorfield==0.7.1
drf-extra-fields==3.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.19.2
